requirements.txt
cortex-code-agent-sdk==1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
snowflake-connector-python>=3.6.0
PyJWT>=2.8.0
mcp>=1.0.0
